templates/ai-native-codex-package-template/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-native-codex-package-template"
version = "0.1.0"
description = "Deterministic template for AI-native packages with benchmark-friendly validation and docs contracts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8,<9",
    "jsonschema>=4.25,<5",
    "PyYAML>=6.0.2,<7",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.18,<2",
    "pytest>=9.0,<10",
    "ruff>=0.12,<0.13",
    "types-PyYAML>=6.0.12,<7",
]

[project.scripts]
ai-native-package = "ai_native_package.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ai_native_package*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
files = ["src"]
python_version = "3.12"
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["jsonschema", "jsonschema.*"]
ignore_missing_imports = true
